Refactor codebase to use float instead of double for consistency; update model training, loss functions, and optimizers; enhance OpenBLAS integration and improve memory management in autograd operations.

+++ include/autograd/Tensor.hpp
@@ -4,7 +4,7 @@
 namespace mlengine::autograd {
 
 using MatrixRM =
-    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>;
+    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>;
 
 struct Tensor {
   MatrixRM data;
